+++ pakyow-core/lib/pakyow/runnable/container.rb
@@ -24,8 +24,11 @@ module Pakyow
     #
     #   * `:threaded` - Each service runs in a thread within the current service.
     #
-    # Forked is the default strategy, falling back to Threaded on platforms that don't support fork.
-    # The strategy can be specified explicitly by passing the `:strategy` runtime option.
+    #   * `:hybrid` - Each service runs in its defined strategy, or the best available default.
+    #
+    # Hybrid is the default strategy, where services run in their defined strategy or the best
+    # available (Forked on supported platforms, falling back to Threaded). The container strategy
+    # can be specified explicitly by passing the `:strategy` runtime option.
     #
     # It's expected that only one container is running per process or thread.
     #
@@ -125,7 +128,7 @@ module Pakyow
 
       # Runs the container with options. Supported options include:
       #
-      #   * strategy: `:forked` or `:threaded` (defaults to `:forked`, falling back to `:threaded` on unsupported platforms)
+      #   * strategy: `:forked`, `:threaded`, or `:hybrid` (defaults to `:hybrid`)
       #   * formation: the formation to run (defaults to `Pakyow::Runnable::Formation.all`)
       #   * parent: the parent service or container that this container is running in
       #
@@ -292,7 +295,7 @@ module Pakyow
       end
 
       private def strategy_option(strategy)
-        strategy || self.class.default_strategy
+        strategy || :hybrid
       end
 
       private def formation_option(formation)
@@ -320,15 +323,6 @@ module Pakyow
           # cases elsewhere, such as loading adapters in pakyow/data.
           #
           raise Pakyow::UnknownContainerStrategy.build(error, strategy: strategy)
-        end
-
-        # @api private
-        def default_strategy
-          if ::Process.respond_to?(:fork)
-            :forked
-          else
-            :threaded
-          end
         end
       end
     end